# ws2812: Fix number of nops for AVR at 8 MHz