--- a/drivers/avr/ws2812.c
+++ b/drivers/avr/ws2812.c
@@ -67,31 +67,25 @@ void ws2812_setleds(LED_TYPE *ledarray, uint16_t number_of_leds) {
 #define w_onecycles (((F_CPU / 1000) * w_onepulse + 500000) / 1000000)
 #define w_totalcycles (((F_CPU / 1000) * w_totalperiod + 500000) / 1000000)
 
-// w1 - nops between rising edge and falling edge - low
+// w1_nops - nops between rising edge and falling edge - low
 #if w_zerocycles >= w_fixedlow
-#    define w1 (w_zerocycles - w_fixedlow)
-#else
-#    define w1 0
-#endif
-
-// w2   nops between fe low and fe high
-#if w_onecycles >= (w_fixedhigh + w1)
-#    define w2 (w_onecycles - w_fixedhigh - w1)
-#else
-#    define w2 0
-#endif
-
-// w3   nops to complete loop
-#if w_totalcycles >= (w_fixedtotal + w1 + w2)
-#    define w3 (w_totalcycles - w_fixedtotal - w1 - w2)
-#else
-#    define w3 0
-#endif
-
-#if w1 > 0
-#    define w1_nops w1
+#    define w1_nops (w_zerocycles - w_fixedlow)
 #else
 #    define w1_nops 0
+#endif
+
+// w2_nops - nops between fe low and fe high
+#if w_onecycles >= (w_fixedhigh + w1_nops)
+#    define w2_nops (w_onecycles - w_fixedhigh - w1_nops)
+#else
+#    define w2_nops 0
+#endif
+
+// w3_nops - nops to complete loop
+#if w_totalcycles >= (w_fixedtotal + w1_nops + w2_nops)
+#    define w3_nops (w_totalcycles - w_fixedtotal - w1_nops - w2_nops)
+#else
+#    define w3_nops 0
 #endif
 
 // The only critical timing parameter is the minimum pulse length of the "0"
@@ -102,18 +96,6 @@ void ws2812_setleds(LED_TYPE *ledarray, uint16_t number_of_leds) {
 #elif w_lowtime > 450
 #    warning "Light_ws2812: The timing is critical and may only work on WS2812B, not on WS2812(S)."
 #    warning "Please consider a higher clockspeed, if possible"
-#endif
-
-#if w2 > 0
-#    define w2_nops w2
-#else
-#    define w2_nops 0
-#endif
-
-#if w3 > 0
-#    define w3_nops w3
-#else
-#    define w3_nops 0
 #endif
 
 #define w_nop1 "nop      \n\t"