## Description

When trying to calculate the number of nops for AVR running at 8 MHz, the value of `w3` is expected to be negative; however, because `F_CPU` is defined in tmk_core/avr.mk with the `UL` suffix, the preprocessor performs its calculations using `unsigned long`, getting a very large positive number instead of the expected negative number; this then results in generating code with a huge number of nops.  Fix the broken calculations by performing a comparison before subtraction, so that the unsigned number wraparound does not occur.

The keyboard which triggers the problem is `handwired/promethium`; the buggy code silently compiles, but the resulting timings would be wrong (but apparently not wrong enough to be noticed, because for most WS8212-like LEDs a low level pulse of 6 us is not enough to trigger a reset).

-----

~~I cannot actually test the change on a real hardware (I tested that it does not break the generated code for XD87 HS, which has ATmega32u4 running at 16 MHz, but I don't have any hardware which uses 8 Mhz).~~ *Update:* Got a 3.3V Pro Micro board for testing and verified that the code actually works (at least with those LEDs that I have).

I looked at the Travis build report for #9558 and found a single keyboard which actually has ATmega32u4 running at 8 MHz with some WS2812-style LEDs connected, then looked at the generated assembly code and noticed that it is obviously wrong (has lots of NOPs), and the fix makes it look right (no NOPs at the end of the loop).

~~It should be possible to simplify the code by removing the subsequent checks like `w1 > 0`, but I did not touch them for now to avoid conflicts with #9558.~~ *Update:* Removed useless code.

## Types of Changes

<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply. -->
- [x] Core
- [x] Bugfix
- [ ] New feature
- [ ] Enhancement/optimization
- [ ] Keyboard (addition or update)
- [ ] Keymap/layout/userspace (addition or update)
- [ ] Documentation

## Checklist

<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] My code follows the code style of this project.
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [x] I have read the [**CONTRIBUTING** document](https://docs.qmk.fm/#/contributing).
- [ ] I have added tests to cover my changes.
- [x] I have tested the changes and verified that they work and don't break anything (as well as I can manage).