@mtei 2020-07-06
Looks good to me. But...

> It should be possible to simplify the code by removing the subsequent checks like w1 > 0, 

I recommend this. After merging this PR, you can fix #9558 .