@sigprof 2021-02-08
Now I actually tested the code on a 3.3V Pro Micro board, and even tried to make some actual measurements of the generated PWM signal (using Input Capture on STM32F411 @ 96 MHz).

Before the fix: T0H = 250 ns, T1H = 875 ns, total period = 5250 us (within a byte) or 6250 us (between bytes).
After the fix: T0H = 250 ns, T1H = 875 ns, total period = 1375 us (within a byte) or 2375 us (between bytes).

I also verified timings on a 5V Pro Micro:
Before the fix: T0H = 312.5 ns, T1H = 875 ns, total period = 1250 us (within a byte) or 1750 us (between bytes).
After the fix: T0H = 312.5 ns, T1H = 875 ns, total period = 1250 us (within a byte) or 1750 us (between bytes) — so the fix did not change anything for the 16 MHz configuration, as expected.